Clean

+++ loadbalancer/src/pt/ulisboa/tecnico/cnv/loadbalancer/sudoku/SudokuRequest.java
@@ -16,7 +16,7 @@ public class SudokuRequest implements Runnable {
     private static final int SUDOKU_REQUEST_SUCCESS = 200;
     private static final double MIN_COST_SCALE = 1.5 * Math.pow(10, -2); //Allow to lose up to ~99% of the cost due to time elapsing
     private static final long REQUEST_COST_LOSS_SLOPE = 10000; //Measured to be less than the average since we prefer to overmeasure instead of undermeasure
-    private static final long MAX_REQUEST_COST = 3 * Math.pow(10, 9); //Maximize to cost to ~ that espected for a 5 minute request, avoiding launching multiple instances for a single gigantic request
+    private static final long MAX_REQUEST_COST = (long) (3 * Math.pow(10, 9)); //Maximize to cost to ~ that espected for a 5 minute request, avoiding launching multiple instances for a single gigantic request
     private final SudokuParameters parameters;
     private final long startingCost;
     private final long minCost;
